requirements.txt
# Packages that can be insalled with 'pip install -r requirements.txt'
imagesoup
mutagen==1.35.1
Pillow==3.4.2
pyinstaller==3.3

# Packages that do not allow pip installation:
#   ffmpeg
#   Python 3.6